Support marshalling of JSObject
Marshaller can now marshal a JSObject value that is an array or
an object. Basically any non-array is treated as an object, because
JSObject doesn't have the relevant methods to do any more than that.

Also NashornDebuggerHost can get properties from a JSObject - it
reads keys and values and returns Data property descriptors.

Fixes #6