ScriptFactory.considerReferenceType is now callback-based
Using a Future was not good, since it meant that the VM would
resume before we could handle the new script and set actual breakpoints
for previously requested breakpoints. The callback solution is better
since it ensures that breakpoint setting happens on the correct thread
and before the VM is resumed.