Changed handling of replaced scripts
A replacement script isn't published as an 'ndx' version. Instead it
truly replaces the old script which can be seen as 'delisted'.

This is particularly important when a script uses a source map, since
the source map may refer to the real name of the script (obviously),
which makes it problematic to change the script URL.

Fixes #98